+++ mayan/apps/linking/forms.py
@@ -23,13 +23,31 @@ class SmartLinkForm(forms.ModelForm):
 
 
 class SmartLinkConditionForm(forms.ModelForm):
+    fieldsets = (
+        (
+            _(message='General'), {
+                'fields': ('enabled', 'inclusion',)
+            },
+        ),
+        (
+            _(message='Foreign document'), {
+                'fields': ('foreign_document_data',)
+            },
+        ),
+        (
+            _(message='Local document'), {
+                'fields': ('operator', 'expression', 'negated')
+            },
+        ),
+    )
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         choices = []
         choices.append(
             (
                 ModelField.class_label, ModelField.get_choices_for(
-                    model=Document,
+                    model=Document
                 )
             )
         )
@@ -45,12 +63,21 @@ class SmartLinkConditionForm(forms.ModelForm):
         self.fields['foreign_document_data'] = form_fields.ChoiceField(
             choices=choices, label=_(message='Foreign document field')
         )
+        self.fields['foreign_document_data'].widget.attrs = {
+            'class': 'form-control select2-templating'
+        }
         self.fields['expression'] = ModelTemplateField(
             initial_help_text=self.fields['expression'].help_text,
             label=self.fields['expression'].label, model=Document,
             model_variable='document', required=False
         )
+        self.fields['inclusion'].widget.attrs = {
+            'class': 'form-control select2'
+        }
+        self.fields['operator'].widget.attrs = {
+            'class': 'form-control select2'
+        }
 
     class Meta:
-        model = SmartLinkCondition
         exclude = ('smart_link',)
+        model = SmartLinkCondition